requirements.txt
python					3.7.0
Django                  3.2.16
django-cors-headers     3.13.0
django-filter           22.1
django-rest-authtpython					3.7.0
Django                  3.2.16
django-cors-headers     3.13.0
django-filter           22.1
django-rest-authtoken   2.1.4
django-rest-swagger     2.2.0
djangorestframework     3.14.0
djangorestframework-jwt 1.11.0
Markdown                3.4.1
pip                     22.3.1
PyJWT                   1.7.1
pytz                    2022.7
requests                2.28.1
setuptools              40.6.2


